apps/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.17.2",
    "arq>=0.26.3",
    "asyncpg>=0.31.0",
    "boto3>=1.35.0",
    "email-validator>=2.3.0",
    "fastapi>=0.128.0",
    "firebase-admin>=7.1.0",
    "google-cloud-aiplatform>=1.132.0",
    "greenlet>=3.3.0",
    "honcho>=2.0.0",
    "httpx>=0.28.1",
    "itsdangerous>=2.2.0",
    "presidio-analyzer>=2.2.360",
    "presidio-anonymizer>=2.2.360",
    "pydantic-settings>=2.12.0",
    "pyotp>=2.9.0",
    "python-ulid>=3.1.0",
    "secure>=1.0.1",
    "sentry-sdk[fastapi]>=2.48.0",
    "slowapi>=0.1.9",
    "sqladmin>=0.22.0",
    "sqlalchemy>=2.0.45",
    "stripe>=14.1.0",
    "structlog>=25.5.0",
    "user-agents>=2.2.0",
    "uvicorn[standard]>=0.40.0",
    "weasyprint>=62.0",
    "jinja2>=3.1.0",
]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "ARG"]
ignore = [
    "B008",   # Do not perform function call `Depends` in argument defaults
    "F821",   # Undefined name (false positives in SQLAlchemy models)
    "ARG001", # Unused function argument
    "ARG002", # Unused method argument
    "E501",   # Line too long
    "E402",   # Module level import not at top of file
    "B904",   # Within an `except` clause, raise exceptions with `raise ... from err`
    "SIM117", # Use a single `with` statement with multiple contexts
    "SIM102", # Use a single `if` statement instead of nested `if` statements
    "UP045",  # Use `X | None` for type annotations
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
asyncio_default_test_loop_scope = "session"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.9.1",
    "pre-commit>=4.0.1",
]
